Add formatCurrency function
Added function that takes an amount as input and formats it with the selected currency symbol. Also updated JS task for future to easily change the formatting logic in one place if needed, without having to modify multiple parts of the JS.

Was having issue with NaN error in summary field. Adjusted code to fix,ensuring that the summary is correctly updated.