More updates to README. Added Favicon
Made further updates to README. Focused on structure as well as more information about testing. This is nearly finished.

I have added a Favicon to the webpage.

Updated JS function to clear income and expenses field when something is submitted. This saves the user having to manually change before adding a new income or expense